run slides in lexicographic order
It's a nusiance to have to shift many slides just to insert one.
Look for slides in lexicographic order, rather than strictly numeric
order, in order to facilitate grouping of slides

Note that this change is NOT BACKWARDS COMPATIBLE.  It may result in
existing slideshows being run in a different order.